Fix JXPath expression for absent properties.
In order to verify the fix, an integration test was added. Prior to the fix,
under certain circumstances, the plugin would warn that dependency resolution
yielded two objects. The fix resolves this, and the integration test verifies
that application of the plugin to a previously problematic POM no longer
triggers a warning.